pyproject.toml
[project]
name = "langchain-course-p"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.11.0",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langchain-tavily>=0.2.11",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.23",
]
